Add node-udp-logger image.
This image will be used to test outging udp traffic in mirrord.
The udp-logger will be deployed such that it is only reachable from
the target pod, and the `mirrord`ed process will send a message to
the logger over UDP. The logger logs the message and the tests read
the logs to verify that the message arrived.